+++ components/modal-assistants/ben/index.ts
@@ -1,3 +1,0 @@
-export { BenTrigger } from "./trigger";
-export { BenDrawer } from "./drawer";
-export { BenIntroBubble } from "./intro-bubble";