+++ components/universal-dialog/index.ts
@@ -1,1 +1,0 @@
-export { UniversalDialog, type DialogContext, type ProjectResult } from "./universal-dialog";